// src/store/actions/reservations/interfaces.ts
export enum getReservationsType {
  GET_RESERVATIONS = 'GET_RESERVATIONS',
  GET_RESERVATIONS_FAIL = 'GET_RESERVATIONS_FAIL',
  GET_RESERVATIONS_SUCCESS = 'GET_RESERVATIONS_SUCCESS',
  SHOW_WARNING = 'SHOW_WARNING',
  CANCEL_DELETE = 'CANCEL_DELETE',
  ADD_RESERVATION_FAIL = 'ADD_RESERVATION_FAIL',
  EDIT_RESERVATION_FAIL = 'EDIT_RESERVATION_FAIL',
  RESET_RESERVATION_ERRORS = 'RESET_RESERVATION_ERRORS',
  GET_CURRENT_RESERVATION = 'GET_CURRENT_RESERVATION',
  ADD_RESERVATION_REQ = "ADD_RESERVATION_REQ",
  EDIT_RESERVATION_REQ = "EDIT_RESERVATION_REQ"
}